+++ app/src/main/java/com/app/arcabyolimpo/data/remote/api/ArcaApi.kt
@@ -19,6 +19,7 @@ import com.app.arcabyolimpo.data.remote.dto.supplies.AcquisitionDto
 import com.app.arcabyolimpo.data.remote.dto.productbatches.ProductBatchDto
 import com.app.arcabyolimpo.data.remote.dto.product.ProductDetailDto
 import com.app.arcabyolimpo.data.remote.dto.product.ProductRegisterInfoDto
+import com.app.arcabyolimpo.data.remote.dto.productbatches.ProductBatchDto
 import com.app.arcabyolimpo.data.remote.dto.productbatches.ProductBatchModifyDto
 import com.app.arcabyolimpo.data.remote.dto.productbatches.ProductBatchRegisterDto
 import com.app.arcabyolimpo.data.remote.dto.supplies.DeleteDto
@@ -32,17 +33,21 @@ import com.app.arcabyolimpo.data.remote.dto.supplies.RegisterSupplyBatchDto
 import com.app.arcabyolimpo.data.remote.dto.supplies.SuccessMessageDto
 import com.app.arcabyolimpo.data.remote.dto.supplies.SuppliesListDto
 import com.app.arcabyolimpo.data.remote.dto.supplies.SupplyBatchDto
+import com.app.arcabyolimpo.data.remote.dto.supplies.SupplyBatchItemDto
+import com.app.arcabyolimpo.data.remote.dto.supplies.SupplyBatchListDto
+import com.app.arcabyolimpo.data.remote.dto.supplies.SupplyBatchOneDto
 import com.app.arcabyolimpo.data.remote.dto.supplies.SupplyDto
 import com.app.arcabyolimpo.data.remote.dto.supplies.WorkshopCategoryListDto
 import com.app.arcabyolimpo.data.remote.dto.user.UserDto
-import com.app.arcabyolimpo.data.remote.dto.user.registeruser.RegisterResponseDto
 import com.app.arcabyolimpo.data.remote.dto.user.registeruser.RegisterUserDto
+import com.app.arcabyolimpo.data.remote.dto.user.updateuser.UpdateUserDto
 import com.app.arcabyolimpo.data.remote.dto.workshops.AddNewWorkshopDto
 import com.app.arcabyolimpo.data.remote.dto.workshops.DeleteResponseWorkshopDto
 import com.app.arcabyolimpo.data.remote.dto.workshops.DeleteWorkshopDto
 import com.app.arcabyolimpo.data.remote.dto.workshops.WorkshopDto
-import com.app.arcabyolimpo.data.remote.dto.workshops.WorkshopsListDto
 import com.app.arcabyolimpo.data.remote.dto.workshops.WorkshopResponseDto
+import com.app.arcabyolimpo.data.remote.dto.workshops.WorkshopsListDto
+import com.app.arcabyolimpo.domain.model.supplies.RegisterSupplyBatch
 import okhttp3.MultipartBody
 import okhttp3.RequestBody
 import retrofit2.Response
@@ -85,8 +90,8 @@ interface ArcaApi {
 
     @POST("user/update")
     suspend fun updateUser(
-        @Body user: UserDto,
-    ): RegisterResponseDto
+        @Body user: UpdateUserDto,
+    ): UserDto
 
     @POST("user/delete/{id}")
     suspend fun deleteUser(
@@ -140,6 +145,11 @@ interface ArcaApi {
     suspend fun getSupplyBatchById(
         @Path("id") id: String,
     ): SupplyBatchDto
+
+    @GET("supplyBatch/modify/{idSupplyBatch}")
+    suspend fun getSupplyBatchOne(
+        @Path("idSupplyBatch") idSupplyBatch: String,
+    ): SupplyBatchOneDto
 
     @POST("supplyBatch/addBatch")
     suspend fun registerSupplyBatch(
@@ -215,8 +225,6 @@ interface ArcaApi {
         @Body requestBody: DeleteWorkshopDto,
         // DeleteResponseDto is for the response, for the snackbar
     ): DeleteResponseWorkshopDto
-
-
 
     // Beneficiary -------------
     @GET("beneficiary/list")
@@ -326,6 +334,7 @@ interface ArcaApi {
         @Part("Disponible") status: RequestBody,
         @Part image: MultipartBody.Part?,
     )
+
     @DELETE("product/{idProduct}")
     suspend fun deleteProduct(
         @Path("idProduct") idProduct: String,
@@ -350,7 +359,6 @@ interface ArcaApi {
         @Query("disponible") disponible: Int,
     ): List<ProductDto>
 
-
     @GET("product/workshop")
     suspend fun getProductsByWorkshop(
         @Query("idTaller") idTaller: String,
@@ -361,6 +369,24 @@ interface ArcaApi {
         @Query("orderBy") orderBy: String,
         @Query("direction") direction: String,
     ): List<ProductDto>
+
+    /**
+     * Modifies a supply batch.
+     *
+     * @param idSupplyBatch The ID of the supply batch to be modified.
+     */
+    @POST("supplyBatch/update/{idSupplyBatch}")
+    suspend fun modifySupplyBatch(
+        @Path("idSupplyBatch") idSupplyBatch: String,
+        @Body batch: RegisterSupplyBatchDto,
+    ): SuccessMessageDto
+
+    @GET("supplyBatch/dates/{date}/{idInsumo}")
+    suspend fun supplyBatchList(
+        @Path("date") expirationDate: String,
+        @Path("idInsumo") idSupply: String,
+    ): List<SupplyBatchItemDto>
+        
     @GET("product/{idProduct}/update")
     suspend fun getProduct(
         @Path("idProduct") idProduct: String,