+++ app/src/main/java/com/app/arcabyolimpo/data/repository/workshops/WorkshopRepositoryImpl.kt
@@ -40,7 +40,7 @@ class WorkshopRepositoryImpl
                 id = dto.id,
                 idUser = dto.idUser,
                 nameWorkshop = dto.name,
-                url = dto.image,
+                url = dto.image.orEmpty(),
                 status = 0,
                 description = "",
                 startHour = "",